Problem Description You are given 3 integer numbers: a, b and c, which will represent the 3 sides of a triangle. Your task is to check whether the triangle is valid.
If it is, print "Triangle is valid.".
Otherwise print "Invalid Triangle." and end the program.

If it is valid, you have to check if it is a right triangle (a2 + b2 == c2).
If it is a right triangle, print "Triangle has a right angle between sides a and b", depending on which side forms a right angle. If the sides b and c form a right angle, print "Triangle has a right angle between sides b and c", and so on.